Dropped fanout exchange, working with single queue

+++ worker/pipeline/component.py
@@ -1,6 +1,5 @@
 """Component"""
 import os
-import urllib.request
 from bson.objectid import ObjectId
 
 
@@ -12,17 +11,16 @@ class Component:
     self.oid = oid
     self.image_url = image_url
     self.mime_type = mime_type
-    # download the image
-    urllib.request.urlretrieve(image_url, f'image-{oid}-{name}')
 
   def clear_files(self):
     """Clears downloaded or any processing files"""
-    os.remove(self.file_name)
+    if os.path.exists(self.file_name):
+      os.remove(self.file_name)
 
   @property
   def file_name(self):
     """Returns the downloaded file name"""
-    return f'image-{self.oid}-{self.name}'
+    return f'image-{self.oid}'
 
   def update(self, data):
     """Updates database with the pipeline result"""
